+++ src/index.ts
@@ -1,0 +1,8 @@
+export { Metrics, MetricsOptions } from './metrics';
+export {
+  ConnectQOS as default,
+  ConnectQOSOptions,
+  ConnectQOSMiddleware,
+  BeforeThrottleFn,
+  GetMiddlewareOptions
+} from './connect';